+++ CHANGELOG.md
@@ -5,11 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.4.0] - 2019-06-29
 ### Added
-- Generates posts and tags archive pages.
 - Date format to config.
 - Site config to template locals.
 - Ability to omit template file extension.
+- Posts and tags archive page generation.
 
 ## [0.3.0] - 2019-06-21
 ### Added